stoned v0.1.0 Stoned.Backoffice.Report.DataFlow

Modulo responsavel por processar eventos para gerar relatorio

Esse modulo constroi um Flow para processar os eventos e gerar um relatorio backoffice. Primariamente ele utiliza Stoned.Backoffice.Report.DataProducer como Source

Functions

generate_keys(date)

Gera as chaves para criar o objeto reposta como relatorio

dia = d_17_03_2020

mes = m_03_2020

ano = y_2020

process_data(event, acc)

Funcao que processa os eventos

Ao receber :empty significa que ja esta no final e apenas retorna o acumulador

Atualiza a chave do acumulador com o valor do evento ou se a chave ja existir faz a soma.

start_flow(start_date)

Cria o flow de processamento de eventos para gerar o Relatorio

Utiliza Flow.from_specs/1 para utilizar um produtor GenStage como Source e inicia-lo como processo filho.