stoned v0.1.0 Stoned.Backoffice.Report.DataProducer

Produtor Genstage para relatorio

Functions

empty_event()

Valor utilizado para indicar o esgotamento dos eventos para serem processados.

meet_demand(events, demand)

Funcao para atender a demanda dos consumidores

Recebe os eventos gerados atraves da consulta em EventsModel.fetch_events/2 e a demanda.

Se identificar que o numero de eventos obtidos foi menor do que a demanda, completa a demanda com :empty_event en entao sinaliza atraves de GenStage.async_info/2 para encerrar o produtor e assim encerrar o Flow associado a ele.

start_link(data)