stoned v0.1.0 Stoned.Banking.Account

Conta do usuario para Depositos, Saques e Transferencias.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

handle_cast(msg, state)

Cast ativado para realizar transferencias

Esse cast visa efetuar o recebimento de uma transferencia

operate(state, type, value)

Funcao que opera com o comportamento basico das operacoes