stoned v0.1.0 Stoned.Banking.AccountAPI

API de acesso a conta do Cliente em Stoned.Banking.Account

Functions

active(pid)

deposit(pid, value)

lookup(email)

Checa o Registry se o usuario ja se encontra com processo ativo.

Caso nao esteja ativo, starta um processo com o DynamicSupervisor.

receive(list)

start(email)

transfer(pid, payload)

withdraw(pid, value)